Refresh style of release details modal
