+++ src/styles/components/_release_selector.sass
@@ -39,8 +39,6 @@
 
 .release-selector-modal--release-details
   border-bottom: 1px solid #eee
-  padding-bottom: 18px
-  margin-bottom: 25px
 
   .progress_button--container
     margin-left: 6px
